@@ -29,17 +29,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.0",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
-    "@angular/service-worker": "^5.2.8",
-    "@fortawesome/angular-fontawesome": "0.1.0-9",
+    "@angular/animations": "^6.0.4",
+    "@angular/common": "^6.0.4",
+    "@angular/compiler": "^6.0.4",
+    "@angular/core": "^6.0.4",
+    "@angular/forms": "^6.0.4",
+    "@angular/http": "^6.0.4",
+    "@angular/platform-browser": "^6.0.4",
+    "@angular/platform-browser-dynamic": "^6.0.4",
+    "@angular/router": "^6.0.4",
+    "@angular/service-worker": "^6.0.4",
+    "@fortawesome/angular-fontawesome": "0.1.0-10",
     "@fortawesome/fontawesome": "^1.1.7",
     "@fortawesome/fontawesome-svg-core": "^1.2.0-11",
     "@fortawesome/free-brands-svg-icons": "^5.1.0-8",
@@ -50,40 +50,42 @@
     "@ngx-share/core": "^5.0.1",
     "angular-bootstrap-md": "^6.0.1",
     "angular-font-awesome": "^3.1.2",
-    "angularfire2": "^5.0.0-rc.6.0",
-    "body-parser": "~1.5.2",
+    "angularfire2": "^5.0.0-rc.7",
+    "body-parser": "~1.18.3",
     "chart.js": "^2.5.0",
     "core-js": "^2.4.1",
-    "express": "~4.7.2",
-    "firebase": "4.12.1",
+    "express": "~4.16.3",
+    "firebase": "5.0.4",
     "font-awesome": "^4.7.0",
     "hammerjs": "^2.0.8",
-    "method-override": "~2.1.2",
-    "rxjs": "^5.5.6",
-    "webpack": "^3.11.0",
-    "zone.js": "^0.8.19"
+    "method-override": "~2.3.10",
+    "rxjs": "^6.2.0",
+    "rxjs-compat": "^6.2.0",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.6.7",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
+    "@angular-devkit/build-angular": "~0.6.8",
+    "@angular/cli": "^6.0.8",
+    "@angular/compiler-cli": "^6.0.4",
+    "@angular/language-service": "^6.0.4",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
+    "@types/node": "~10.3.1",
     "codelyzer": "^4.0.1",
-    "jasmine-core": "~2.8.0",
+    "jasmine-core": "~3.1.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~2.0.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
+    "karma-coverage-istanbul-reporter": "^2.0.1",
     "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-jasmine-html-reporter": "^1.1.0",
     "karma-phantomjs-launcher": "^1.0.4",
     "mongodb-stitch": "^3.0.0",
-    "protractor": "~5.1.2",
+    "protractor": "~5.3.2",
     "sw-precache-webpack-plugin": "^0.11.5",
-    "ts-node": "~4.1.0",
-    "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "ts-node": "~6.1.0",
+    "tslint": "~5.10.0",
+    "typescript": "^2.7.2",
+    "webpack": "^4.12.0"
   }
 }
