@@ -1,12 +1,13 @@
 import { Injectable } from '@angular/core';
 import { CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot, Router } from '@angular/router';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { AuthService } from '@wh-share/auth.service';
 import * as _ from 'lodash';
+import { take, map, tap } from 'rxjs/operators';
 
-import 'rxjs/add/operator/do';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/take';
+
+
+
 
 @Injectable()
 export class AuthGuard implements CanActivate {
@@ -14,16 +15,17 @@ export class AuthGuard implements CanActivate {
   constructor(private auth: AuthService, private router: Router) { }
 
   canActivate(
-    next: ActivatedRouteSnapshot,
-    state: RouterStateSnapshot): Observable<boolean> | Promise<boolean> | boolean {
+  ): Observable<boolean> | Promise<boolean> | boolean {
     return this.auth.user
-      .take(1)
-      .map(user => _.has(_.get(user, 'roles'), 'admin' || 'editor'))
-      .do(authorized => {
-        if (!authorized) {
-          console.log('route prevented!');
-          this.router.navigate(['/login']);
-        }
-      });
+      .pipe(
+        take(1),
+        map(user => _.has(_.get(user, 'roles'), 'admin' || 'editor')),
+        tap(authorized => {
+          if (!authorized) {
+            console.log('route prevented!');
+            this.router.navigate(['/login']);
+          }
+        })
+      );
   }
 }
