@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { Attachment } from '@wh-objects/attachment';
 import { HttpClient } from '@angular/common/http';
 
