@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { HttpClient } from '@angular/common/http';
 import { Ligue, Statistik, Club, StatGame, StatGoals } from '@wh-objects/hvw';
 import { teamStatistics } from '@wh-objects/team-statistics';
