@@ -1,7 +1,6 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { Observable } from 'rxjs/Observable';
-import { ReplaySubject } from 'rxjs/ReplaySubject';
+import { Observable ,  ReplaySubject } from 'rxjs';
 import { Sponsors } from '@wh-objects/sponsors';
 
 const apiSponsorImages = "https://wp.willsbach-handball.de/wp-json/wp/v2/media?_embed&search=sponsors";
